Fixed serving Journal Entries, added server components to use getPayloadHMR
